requirements.txt
asgiref==3.11.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
Django==5.2.11
django-cors-headers==4.9.0
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework-stubs==3.16.8
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
google-auth==2.48.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
pillow==12.1.1
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
sqlparse==0.5.5
tomli==2.4.0
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
